Did some minor cleanup
Closes #1
