+++ LJ-IAM-WebService.sln
@@ -21,7 +21,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Diagnostics", "src\Diagnost
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Domain", "src\Domain\Domain.csproj", "{1BC36B7C-D6CB-40F5-A6BC-B1727184F41D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common", "src\Application.Contracts\Common.csproj", "{A4E3A05D-6346-42A6-8682-13B4113C13C4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Common", "src\Application.Contracts\Common.csproj", "{A4E3A05D-6346-42A6-8682-13B4113C13C4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
