--- src/Diagnostics/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Diagnostics;
-
-public class Class1
-{
-
-}
