+++ src/Diagnostics/DiagnosticsConfig.cs
@@ -1,0 +1,6 @@
+﻿namespace Diagnostics;
+
+public static class DiagnosticsConfig
+{
+    public const string ServiceName = "LF-IAM-WebService";
+}
