--- src/Infrastructure/Abstractions/Validations/IBearerTokenUsageTypeValidation.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Infrastructure.Validations.BearerTokenUsageType;
-
-namespace Infrastructure.Abstractions.Validations;
-
-public interface IBearerTokenUsageTypeValidation
-{
-    Task<BearerTokenUsageTypeValidationResponse> ValidateAsync();
-}
