+++ src/Infrastructure/Services/AuthorizeResultService.cs
@@ -1,17 +1,15 @@
 ﻿using System.IdentityModel.Tokens.Jwt;
-using System.Linq;
 using System.Security.Claims;
 using System.Security.Cryptography;
 using System.Text;
 using Abstractions;
-using Application.Abstractions.Services;
 using Common;
 using Domain.Modules.Clients;
 using Domain.Modules.Devices;
 using Domain.Modules.OAuth;
-using Infrastructure.Abstractions.Services;
 using Infrastructure.Enumerations;
 using Infrastructure.Options;
+using Infrastructure.Services.Interfaces;
 using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.IdentityModel.Tokens;
