+++ src/Infrastructure/Services/Interfaces/IDeviceAuthorizationService.cs
@@ -1,7 +1,7 @@
 ﻿using Common.Devices;
 using Microsoft.AspNetCore.Http;
 
-namespace Infrastructure.Abstractions.Services;
+namespace Infrastructure.Services.Interfaces;
 
 public interface IDeviceAuthorizationService
 {
