+++ src/Infrastructure/Services/Interfaces/IUserManagerService.cs
@@ -1,7 +1,7 @@
 ﻿using Common;
 using Domain.Modules.Users;
 
-namespace Infrastructure.Abstractions.Services;
+namespace Infrastructure.Services.Interfaces;
 
 public interface IUserManagerService
 {
