+++ src/Infrastructure/Services/Interfaces/IUserService.cs
@@ -1,4 +1,4 @@
-﻿using System.Security.Claims;
+﻿using Abstractions;
 using Common;
 using Domain.Modules.Users;
