+++ src/Infrastructure/Services/Interfaces/LdapService.cs
