+++ src/Infrastructure/Services/Responses/DeviceAuthorizationResponse.cs
