+++ src/Infrastructure/Services/Responses/UserInfoResponse.cs
