+++ src/Infrastructure/Validations/BearerTokenUsageType/IBearerTokenUsageTypeValidation.cs
@@ -1,0 +1,6 @@
+﻿namespace Infrastructure.Validations.BearerTokenUsageType;
+
+public interface IBearerTokenUsageTypeValidation
+{
+    Task<BearerTokenUsageTypeValidationResponse> ValidateAsync();
+}
