+++ src/Infrastructure/Validations/DeviceAuthorization/IDeviceAuthorizationValidation.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.Http;
+
+namespace Infrastructure.Validations.DeviceAuthorization;
+
+public interface IDeviceAuthorizationValidation
+{
+    Task<DeviceAuthorizationValidationResponse> ValidateAsync(HttpContext httpContext);
+}
