+++ src/Infrastructure/Validations/TokenRevocation/ITokenRevocationValidation.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.Http;
+
+namespace Infrastructure.Validations.TokenRevocation;
+
+public interface ITokenRevocationValidation
+{
+    Task<TokenRevocationValidationResponse> ValidateAsync(HttpContext context);
+}
