+++ src/Persistence/DbContexts/ApplicationDbContext.cs
@@ -2,6 +2,7 @@
 using Domain.Modules.OAuth;
 using Domain.Modules.Users;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+using Domain.Modules.Devices;
 
 namespace Persistence.DbContexts;
 
@@ -11,6 +12,8 @@ public sealed class ApplicationDbContext : IdentityDbContext<User>
     {
     }
 
+    public DbSet<OAuthApplication> OAuthApplications { get; set; }
     public DbSet<OAuthToken> OAuthTokens { get; set; }
+    public DbSet<DeviceFlow> DeviceFlows { get; set; }
 }
