+++ src/WebApi/Controllers/AccountsController.cs
@@ -1,5 +1,5 @@
 ﻿using Common;
-using Infrastructure.Abstractions.Services;
+using Infrastructure.Services.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 
 namespace WebApi.Controllers;
