+++ src/WebApi/Controllers/AuthController.cs
@@ -1,6 +1,6 @@
 ﻿using Application.Abstractions.Services;
 using Common;
-using Infrastructure.Abstractions.Services;
+using Infrastructure.Services.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 
 namespace WebApi.Controllers;
