+++ src/WebApi/Controllers/IntrospectionsController.cs
@@ -1,5 +1,5 @@
 ﻿using Common.Tokens;
-using Infrastructure.Abstractions.Services;
+using Infrastructure.Services.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 
 namespace WebApi.Controllers;
