+++ src/WebApi/Program.cs
@@ -45,9 +45,7 @@ app.UseHttpsRedirection();
 
 app.UseAuthentication();
 app.UseAuthorization();
-//app.UseSession();
 app.MapControllers();
 
-//app.MapIdentityApi<User>();
 
 app.Run();
