+++ src/WebApi/Properties/launchSettings.json
@@ -1,6 +1,6 @@
 ﻿{
   "profiles": {
-    "https": {
+    "Run (HTTPS)": {
       "commandName": "Project",
       "dotnetRunMessages": true,
       "launchBrowser": true,
